+++ libs/sub_client/client_sdl/Car.h
@@ -10,20 +10,20 @@
 
 class Car {
 private:
-    uint16_t x;
-    uint16_t y;
+    float x;
+    float y;
     uint16_t sign;
-    uint16_t angle;
+    float angle;
 
 public:
-    Car(uint16_t &id, uint16_t x, uint16_t y, uint16_t sign, uint16_t angle) : x(x), y(y), sign(sign), angle(angle) {};
+    Car(uint16_t &id, float x, float y, uint16_t sign, float angle) : x(x), y(y), sign(sign), angle(angle) {};
     ~Car() = default;
 
     //getters
-    uint16_t getX() const { return x; };
-    uint16_t getY() const { return y; };
+    float getX() const { return x; };
+    float getY() const { return y; };
     uint16_t getSign() const { return sign; };
-    uint16_t getAngle() const { return angle; };
+    float getAngle() const { return angle; };
 };
